remove old details module
