+++ convex/_generated/api.d.ts
@@ -13,7 +13,6 @@ import type {
   FilterApi,
   FunctionReference,
 } from "convex/server";
-import type * as details from "../details.js";
 import type * as features from "../features.js";
 import type * as imageMigration from "../imageMigration.js";
 import type * as internal_ from "../internal.js";
@@ -29,7 +28,6 @@ import type * as projects from "../projects.js";
  * ```
  */
 declare const fullApi: ApiFromModules<{
-  details: typeof details;
   features: typeof features;
   imageMigration: typeof imageMigration;
   internal: typeof internal_;
